requirements.txt
requests~=2.23.0
django~=2.2.9
pytest~=5.4.3
pytest-django
djangorestframework~=3.12.1
djangorestframework-simplejwt

django-filter~=2.4.0
py~=1.9.0
PyJWT~=1.7.1
cryptography~=2.9.2
pip~=20.0.2
keyring~=18.0.1
attrs~=19.3.0
pytz~=2019.3
toml~=0.10.1
sqlparse~=0.3.1
PyYAML~=5.3.1
wcwidth~=0.2.5
pluggy~=0.13.1
iniconfig~=1.0.1
pexpect~=4.6.0
certifi~=2020.4.5.1
chardet~=3.0.4
urllib3~=1.25.9
idna~=2.9
simplejson~=3.16.0
pyparsing~=2.4.7
six~=1.15.0
setuptools~=45.2.0